backend/agents/lead_qualification/tools.py
from typing import Dict, Optional
import requests
from bs4 import BeautifulSoup
import re

class LeadQualificationTools:
    def __init__(self):
        self.headers = {
            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36'
        }

    async def linkedin_profile_scraper(self, url: str) -> Dict:
        """Scrape LinkedIn profile data"""
        try:
            # Basic implementation - en producción usar LinkedIn API o herramientas más robustas
            response = requests.get(url, headers=self.headers)
            soup = BeautifulSoup(response.content, 'html.parser')

            # Extract basic info
            name = self._extract_name(soup)
            title = self._extract_title(soup)
            company = self._extract_company(soup)
            location = self._extract_location(soup)

            return {
                "name": name,
                "title": title,
                "company": company,
                "location": location,
                "profile_url": url
            }
        except Exception as e:
            return {"error": f"Failed to scrape LinkedIn profile: {str(e)}"}

    async def company_research(self, domain: str) -> Dict:
        """Research company information"""
        try:
            # Implementation using multiple data sources
            company_info = await self._get_company_basic_info(domain)
            employee_count = await self._estimate_company_size(domain)
            industry_info = await self._get_industry_classification(domain)

            return {
                "domain": domain,
                "employee_count": employee_count,
                "industry": industry_info,
                **company_info
            }
        except Exception as e:
            return {"error": f"Failed to research company: {str(e)}"}

    def _extract_name(self, soup) -> str:
        """Extract name from LinkedIn profile soup"""
        name_tag = soup.find('h1')
        return name_tag.get_text(strip=True) if name_tag else "N/A"

    def _extract_title(self, soup) -> str:
        """Extract title from LinkedIn profile soup"""
        title_tag = soup.find('div', class_=re.compile('top-card-layout__entity-info-container'))
        if title_tag:
            title_element = title_tag.find('div', class_=re.compile('top-card-layout__entity-info-container'))
            if title_element:
                 # This is a common pattern, but LinkedIn's class names are dynamic.
                 # A more robust solution would involve inspecting the HTML structure.
                 # For now, a simplified approach.
                 # Attempt to find a common title pattern
                 title_match = re.search(r'title:\s*"([^"]+)"', str(soup))
                 if title_match:
                     return title_match.group(1)
        return "N/A"

    def _extract_company(self, soup) -> str:
        """Extract company from LinkedIn profile soup"""
        company_tag = soup.find('a', class_=re.compile('top-card-link'))
        if company_tag:
             # Similar to title, class names are dynamic.
             # Look for common patterns or data attributes.
             company_match = re.search(r'company:\s*"([^"]+)"', str(soup))
             if company_match:
                 return company_match.group(1)
        return "N/A"

    def _extract_location(self, soup) -> str:
        """Extract location from LinkedIn profile soup"""
        location_tag = soup.find('span', class_=re.compile('top-card-layout__entity-info-container'))
        if location_tag:
             # Look for common patterns or data attributes.
             location_match = re.search(r'location:\s*"([^"]+)"', str(soup))
             if location_match:
                 return location_match.group(1)
        return "N/A"

    # Company research helper methods (Simulated Implementation)
    async def _get_company_basic_info(self, domain: str) -> Dict:
        """Simulated: Get basic info for company"""
        print(f"Simulating getting basic info for company: {domain}")
        # In a real implementation, this would call an external API or scrape
        simulated_data = {
            "example.com": {"company_name": "Example Corp", "industry": "Technology", "description": "A leading tech company."},
            "anothercorp.com": {"company_name": "Another Corp", "industry": "Consulting", "description": "Business consulting services."}
        }
        return simulated_data.get(domain, {"company_name": "Unknown Company", "industry": "Unknown", "description": ""})

    async def _estimate_company_size(self, domain: str) -> Optional[int]:
        """Simulated: Estimate company size"""
        print(f"Simulating estimating company size for: {domain}")
        # In a real implementation, this would call an external API or scrape
        simulated_sizes = {
            "example.com": 300,
            "anothercorp.com": 150
        }
        return simulated_sizes.get(domain, 100) # Default simulated size

    async def _get_industry_classification(self, domain: str) -> Optional[str]:
        """Simulated: Get industry classification"""
        print(f"Simulating getting industry classification for: {domain}")
        # In a real implementation, this would call an external API or scrape
        simulated_industries = {
            "example.com": "Software Development",
            "anothercorp.com": "Management Consulting"
        }
        return simulated_industries.get(domain, "Other")

    async def crm_update(self, lead_data: dict) -> Dict:
        """Simulated: CRM update logic"""
        print(f"Simulating CRM update for lead: {lead_data.get('email')}")
        # TODO: Implement actual CRM integration (e.g., HubSpot API, Salesforce API)
        # This would involve calling the CRM API with lead_data
        # For now, just print the data that would be sent
        print("Simulated CRM update data:")
        print(json.dumps(lead_data, indent=2))
        print("CRM update simulated successfully.")
        return {"status": "success", "message": "CRM update simulated"}

    async def qualification_score(self, profile: dict, icp: dict) -> Dict:
        """Simulated: Qualification scoring logic based on ICP and profile data"""
        print("Simulating qualification scoring based on ICP and profile data...")
        # TODO: Implement actual scoring logic based on ICP and profile data
        # This simulated logic will consider the criteria from the agent's system prompt:
        # - Company Size: 10-500 employees (weight: 30%)
        # - Budget Authority: C-level, VP, Director roles (weight: 25%)
        # - Pain Points: Manual processes, scaling challenges (weight: 25%)
        # - Engagement Level: Active on LinkedIn, posts about challenges (weight: 20%)

        # Simulate profile data points relevant to scoring
        simulated_company_size = profile.get("employee_count", 0)
        simulated_title = profile.get("title", "")
        simulated_pain_points = profile.get("pain_points", []) # Assuming pain points are identified elsewhere
        simulated_engagement_level = profile.get("engagement_level", "low") # Assuming engagement level is identified elsewhere

        score = 0
        reasoning = []

        # Simulate scoring based on criteria (simplified)
        # Company Size (10-500 employees)
        if 10 <= simulated_company_size <= 500:
            score += 30
            reasoning.append(f"Company size ({simulated_company_size}) is within target range (10-500).")
        else:
            reasoning.append(f"Company size ({simulated_company_size}) is outside target range (10-500).")

        # Budget Authority (C-level, VP, Director)
        if any(role in simulated_title.lower() for role in ["c-level", "vp", "director"]):
            score += 25
            reasoning.append(f"Title ({simulated_title}) indicates budget authority.")
        else:
            reasoning.append(f"Title ({simulated_title}) does not clearly indicate budget authority.")

        # Pain Points (Manual processes, scaling challenges) - Simulated check
        if any(pain in " ".join(simulated_pain_points).lower() for pain in ["manual process", "scaling challenge"]):
             score += 25
             reasoning.append("Identified relevant pain points.")
        else:
             reasoning.append("No clear relevant pain points identified.")


        # Engagement Level (Active on LinkedIn, posts about challenges) - Simulated check
        if simulated_engagement_level == "high":
            score += 20
            reasoning.append("High engagement level detected.")
        elif simulated_engagement_level == "medium":
            score += 10
            reasoning.append("Medium engagement level detected.")
        else:
            reasoning.append("Low engagement level detected.")

        # Ensure score is within a reasonable range (0-100)
        score = max(0, min(100, score))

        # Determine recommendation based on score
        recommendation = "pursue" if score >= 70 else ("nurture" if score >= 60 else "disqualify")

        print(f"Simulated score: {score}, Recommendation: {recommendation}")
        return {"score": score, "recommendation": recommendation, "reasoning": ". ".join(reasoning)}
