requirements.txt
swarms==5.5.0
openai==1.30.1
anthropic==0.18.1
google-generativeai==0.4.0
supabase==2.4.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
redis==5.0.1
python-dotenv==1.0.1
pydantic==2.8.2
sqlalchemy==2.0.25
alembic==1.13.1
langchain-community
langchain-experimental
langsmith
sentry-sdk[fastapi]==1.40.0
prometheus-client==0.19.0
